Split corresponding test files
